Pull app booting code into bouncer.app module
This will allow us to import `bouncer` without requiring any other
Python packages to be installed, which we then can use on Jenkins for
getting the bouncer version.
