+++ conf/supervisord.conf
@@ -5,7 +5,7 @@ logfile=/dev/null
 logfile_maxbytes=0
 
 [program:web]
-command=gunicorn -b 0.0.0.0:8000 bouncer:app()
+command=gunicorn -b 0.0.0.0:8000 bouncer.app:app()
 stdout_logfile=NONE
 stderr_logfile=NONE
 stdout_events_enabled=true
